+++ api.js
@@ -1,7 +1,10 @@
 // This is the file you should include in your chat scripts
-// var API = require("./api.js");
+// var API = require("../api.js");
 // (or whatever name you want)
 // See documentation (I'll probably forget to write this, so remind me if I haven't done it yet lol)
+
+/// ALSO this will probably change a lot lol so don't rely on it too much gosh
+
 var User = require("./user.js");
 var Room = require("./room.js");
 var Axios = require("axios");
@@ -15,7 +18,6 @@ var API = exports;
 // <message unfinished>
 API.messageEvents = []; // called before displaying a message
 API.submitEvents = []; // called when input is submitted
-
 
 // maybe don't modify this directly
 API.commands = {}; // commands (key is command name without prefix)
@@ -77,110 +79,20 @@ API.setNicknameHandler = function(func){
 	User.prototype.getNickname = func;
 };
 
-////////// TEMP CHATJS SCRIPTS //////////
-// todo: maybe define nickname getter as a sort of plug-in to Users
-// rather than it being hardcoded
+// todo: maybe wait to load chatjs until
+// for example, User.me is set, etc.
 
-API.addCommand("nick", "set your nickname", function(params) {
-	// The nickname protocol is a bit weird, because of backwards compatibility etc.
-	API.submitMessage("[rpl29nick] "+API.User.me.username+"'s name is now "+params);
-	// the nickname system uses a weird protocol uhhhh
-	// yeah can't wait for rp30 lol
-	function write_persistent(name, value) {
-		
-		function escape_name(name) {
-			// why does this even exist lol I don't remember
-			// wait I think it was for like
-			// utf8?
-			var out = ""
-			for(var i=0; i<name.length; i++){
-				var chr = name.charAt(i);
-				if(chr=="\0" || chr>="\x7F" || chr=="%")
-					out += escape(chr);
-				else
-					out += chr;
-			}
-			return out;
+// Load chatjs
+var pathJoin = require("path").join;
+var chatjsfolder = pathJoin(__dirname, "chatjs");
+var Fs = require("fs");
+Fs.readdirSync(chatjsfolder).forEach(function(file){
+	if (/\.js$/.test(file)) {
+		try {
+			require(pathJoin(chatjsfolder, file));
+		} catch (e) {
+			console.error("Error while loading chatjs script '"+file+"':"+"\n"+e.stack);
+			//API.displayMessage.warning("Error while loading chatjs script '"+file+"':"+"\n"+e.stack);
 		}
-		function encodeBase64(string) {
-			return Buffer.from(String(string), "binary").toString("base64");
-		}
-		API.Axios.get(
-			"https://smilebasicsource.com/query/submit/varstore"+
-				"?nameb64="+encodeBase64(name)+
-				"&valueb64="+encodeBase64(escape_name(value))+
-				"&session="+API.polyChat.session
-		);
-	}
-	write_persistent("nickname_tcf", params || "\r\n");
-});
-
-API.messageEvents.push(function(msg){
-	if (msg.type=="message" && msg.encoding=="text") {
-		var match = msg.text.match(/^\[rpl[23]\dnick] (.*?)'s name is now (.*)/);
-		if (match) {
-			msg.sender.nickname = match[2] || false;
-		}
-	}
-});
-
-API.setNicknameHandler(function(callback){
-	var gotNickname = (success)=>{
-		if (success)
-			this.nicknameCallbacks.forEach(x=>{
-				if (x)
-					x(this);
-			});
-		this.requestedNickname = false;
-		this.nicknameCallbacks = [];
-	}
-	// limit an action to once every `interval` milliseconds
-	var last = 0;
-	function limit(interval, callback) {
-		var now = Date.now();
-		var ok = last + interval;
-		if (now >= ok){
-			last = now;
-			callback();
-		} else {
-			last = ok;
-			setTimeout(callback, ok-now);
-		}
-	}
-	if (this.nickname === undefined) {
-		this.nicknameCallbacks.push(callback);
-		if (!this.requestedNickname) {
-			this.requestedNickname = true;
-			limit(100, ()=>{
-				if (this.nickname !== undefined) {
-					// got nickname while waiting
-					gotNickname(true);
-				} else {
-					console.log("requesting nickname for "+this.username);
-					API.Axios.get("http://smilebasicsource.com/query/tinycomputerprograms?username="+this.username+"&program=nickname").then(response=>{
-						var nickname = unescape(response.data).replace(/\x1B/g,"");
-						if (nickname >= " " && nickname != "\r\n") {
-							this.nickname = nickname;
-							gotNickname(true);
-						} else {
-							this.nickname = false;
-							gotNickname(false);
-							// wait should this still activate the callbacks...
-						}
-					}).catch(error=>{
-						console.error("Nickname request failed");
-						console.error(error);
-						this.nickname = false;
-						gotNickname(false);
-					});
-				}
-			});
-		}
-		return null;
-	} else if (this.nickname === false) {
-		// user doesn't have a nickname set
-		return this.username;
-	} else {
-		return this.nickname;
 	}
 });
